style: better class location
