+++ common-lib/src/main/kotlin/com/github/heheteam/commonlib/database/DatabaseCourseTokenStorage.kt
@@ -1,20 +1,14 @@
 package com.github.heheteam.commonlib.database
 
-import com.github.heheteam.commonlib.Course
 import com.github.heheteam.commonlib.EduPlatformError
 import com.github.heheteam.commonlib.database.table.CourseTokenTable
 import com.github.heheteam.commonlib.interfaces.CourseId
-import com.github.heheteam.commonlib.interfaces.CourseStorage
-import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.interfaces.TokenError
 import com.github.heheteam.commonlib.interfaces.toCourseId
 import com.github.heheteam.commonlib.util.catchingTransaction
-import com.github.heheteam.commonlib.util.raiseError
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
-import com.github.michaelbull.result.binding
-import java.util.*
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
@@ -22,52 +16,6 @@ import org.jetbrains.exposed.sql.deleteWhere
 import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
-
-class CourseTokenService
-internal constructor(
-  private val databaseCourseTokenStorage: DatabaseCourseTokenStorage,
-  private val courseStorage: CourseStorage,
-) {
-  fun createToken(courseId: CourseId): String {
-    val token = UUID.randomUUID().toString()
-    databaseCourseTokenStorage.storeToken(courseId, token)
-    return token
-  }
-
-  fun regenerateToken(courseId: CourseId): String {
-    val newToken = UUID.randomUUID().toString()
-    databaseCourseTokenStorage.setNewToken(courseId, newToken)
-    return newToken
-  }
-
-  fun getCourseIdByToken(token: String): Result<CourseId, TokenError> =
-    databaseCourseTokenStorage.getCourseIdByToken(token)
-
-  fun useToken(token: String): Result<Unit, EduPlatformError> = binding {
-    val tokenExists = databaseCourseTokenStorage.doesTokenExist(token).bind()
-    if (!tokenExists) {
-      raiseError(TokenError.TokenNotFound)
-    } else {
-      Unit
-    }
-  }
-
-  fun registerStudentForToken(
-    studentId: StudentId,
-    token: String,
-  ): Result<Course, EduPlatformError> = binding {
-    val courseIdResult = getCourseIdByToken(token).bind()
-    courseStorage.addStudentToCourse(studentId, courseIdResult).bind()
-    useToken(token).bind()
-    courseStorage.resolveCourse(courseIdResult).bind()
-  }
-
-  fun deleteToken(token: String): Result<Unit, EduPlatformError> =
-    databaseCourseTokenStorage.deleteToken(token)
-
-  fun getTokenForCourse(courseId: CourseId): String? =
-    databaseCourseTokenStorage.getTokenForCourse(courseId)
-}
 
 class DatabaseCourseTokenStorage(val database: Database) {
   init {
